projects/adxrs290: use the new example infrastructure
Make the necessary changes to use the new example infrastructure.
