+++ projects/adxrs290-pmdz/Makefile
@@ -1,9 +1,7 @@
-# Select the example you want to enable by choosing y for enabling and n for disabling
-IIO_EXAMPLE = n
-IIO_TRIGGER_EXAMPLE = n
-IIO_TIMER_TRIGGER_EXAMPLE = y
-
 include ../../tools/scripts/generic_variables.mk
+
+EXAMPLE := iio_example
+include ../../tools/scripts/examples.mk
 
 include src.mk
