+++ projects/adxrs290-pmdz/src/common/common_data.c
@@ -76,52 +76,3 @@ struct adxrs290_init_param adxrs290_ip = {
 	.lpf = ADXRS290_LPF_480HZ,
 	.hpf = ADXRS290_HPF_ALL_PASS
 };
-
-#ifdef IIO_TRIGGER_EXAMPLE
-struct no_os_irq_init_param adxrs290_gpio_irq_ip = {
-	.irq_ctrl_id = GPIO_IRQ_ID,
-	.platform_ops = GPIO_IRQ_OPS,
-	.extra = GPIO_IRQ_EXTRA,
-};
-
-struct iio_hw_trig_init_param adxrs290_gpio_trig_ip = {
-	.irq_id = ADXRS290_GPIO_TRIG_IRQ_ID,
-	.irq_trig_lvl = NO_OS_IRQ_LEVEL_HIGH,
-	.cb_info = {
-		.event = NO_OS_EVT_GPIO,
-		.peripheral = NO_OS_GPIO_IRQ,
-		.handle = ADXRS290_GPIO_CB_HANDLE,
-	},
-	.name = ADXRS290_GPIO_TRIG_NAME,
-};
-#endif
-
-#ifdef IIO_TIMER_TRIGGER_EXAMPLE
-/* ADXRS290 timer init parameter */
-struct no_os_timer_init_param adxrs290_tip = {
-	.id = ADXRS290_TIMER_DEVICE_ID,
-	.freq_hz = ADXRS290_TIMER_FREQ_HZ,
-	.ticks_count = ADXRS290_TIMER_TICKS_COUNT,
-	.platform_ops = TIMER_OPS,
-	.extra = ADXRS290_TIMER_EXTRA,
-};
-
-/* ADXRS290 timer irq init parameter */
-struct no_os_irq_init_param adxrs290_timer_irq_ip = {
-	.irq_ctrl_id = 0,
-	.platform_ops = TIMER_IRQ_OPS,
-	.extra = ADADXRS290_TIMER_IRQ_EXTRA,
-};
-
-/* ADXRS290 timer trigger init parameter */
-struct iio_hw_trig_init_param adxrs290_timer_trig_ip = {
-	.irq_id = ADXRS290_TIMER_TRIG_IRQ_ID,
-	.cb_info = {
-		.event = NO_OS_EVT_TIM_ELAPSED,
-		.peripheral = NO_OS_TIM_IRQ,
-		.handle = ADXRS290_TIMER_CB_HANDLE,
-	},
-	.name = ADXRS290_TIMER_TRIG_NAME,
-};
-
-#endif
