+++ projects/adxrs290-pmdz/src/examples/iio_timer_trigger_example/example.mk
@@ -1,0 +1,5 @@
+IIOD=y
+
+INCS += $(DRIVERS)/gyro/adxrs290/iio_adxrs290.h
+SRCS += $(DRIVERS)/gyro/adxrs290/iio_adxrs290.c \
+	$(DRIVERS)/gyro/adxrs290/iio_adxrs290_trig.c
