+++ projects/adxrs290-pmdz/src/platform/aducm3029/main.c
@@ -34,20 +34,10 @@
 /******************************************************************************/
 /***************************** Include Files **********************************/
 /******************************************************************************/
-#include "platform_includes.h"
+#include "parameters.h"
 #include "common_data.h"
 
-#ifdef IIO_EXAMPLE
-#include "iio_example.h"
-#endif
-
-#ifdef IIO_TRIGGER_EXAMPLE
-#include "iio_trigger_example.h"
-#endif
-
-#ifdef IIO_TIMER_TRIGGER_EXAMPLE
-#include "iio_timer_trigger_example.h"
-#endif
+extern int example_main();
 
 /***************************************************************************//**
  * @brief Main function execution for STM32 platform.
@@ -61,32 +51,7 @@ int main()
 
 	ret = platform_init();
 	if (ret)
-		goto error;
+		return ret;
 
-#ifdef IIO_EXAMPLE
-	ret = iio_example_main();
-	if (ret)
-		goto error;
-#endif
-
-#ifdef IIO_TRIGGER_EXAMPLE
-	ret = iio_trigger_example_main();
-	if (ret)
-		goto error;
-#endif
-
-#ifdef IIO_TIMER_TRIGGER_EXAMPLE
-	ret = iio_timer_trigger_example_main();
-	if (ret)
-		goto error;
-#endif
-
-#if (IIO_EXAMPLE + IIO_TRIGGER_EXAMPLE + IIO_TIMER_TRIGGER_EXAMPLE != 1)
-#error Selected example projects cannot be enabled at the same time. \
-Please enable only one example and rebuild the project.
-#endif
-
-error:
-	return 0;
+	return example_main();
 }
-
